Move booleasn user defaults viewmodel
Move the viewmodel to a separate file and the entire feature to a separate group folder
