+++ WordPress/Classes/ViewRelated/Me/App Settings/Boolean User Defaults/BooleanUserDefaultsDebugViewModel.swift
@@ -1,38 +1,6 @@
 import SwiftUI
 
-private typealias BooleanUserDefaultsSections = [String: BooleanUserDefaults]
-private typealias BooleanUserDefaults = [String: BooleanUserDefault]
-
-struct BooleanUserDefaultsDebugView: View {
-    @StateObject private var viewModel = BooleanUserDefaultsDebugViewModel()
-
-    var body: some View {
-        List {
-            ForEach(viewModel.userDefaultsSections.keys.sorted(), id: \.self) { sectionKey in
-                let userDefaultsSection = viewModel.userDefaultsSections[sectionKey] ?? BooleanUserDefaults()
-
-                Section(header: Text(sectionKey)
-                    .font(.caption)) {
-                        ForEach(userDefaultsSection.keys.sorted(), id: \.self) { userDefaultKey in
-                            let isOn = Binding<Bool>(
-                                get: { userDefaultsSection[userDefaultKey]?.value ?? false },
-                                set: { newValue in viewModel.updateUserDefault(newValue, forSection: sectionKey, forUserDefault: userDefaultKey) }
-                            )
-                            Toggle(userDefaultsSection[userDefaultKey]!.title, isOn: isOn)
-                                .font(.caption)
-                        }
-                    }
-            }
-        }
-        .navigationTitle(Strings.title)
-        .searchable(text: $viewModel.searchQuery, placement: .navigationBarDrawer(displayMode: .always))
-        .onAppear {
-            viewModel.load()
-        }
-    }
-}
-
-private final class BooleanUserDefaultsDebugViewModel: ObservableObject {
+final class BooleanUserDefaultsDebugViewModel: ObservableObject {
     @Published private var allUserDefaultsSections = BooleanUserDefaultsSections()
     @Published var searchQuery: String = ""
 
@@ -112,12 +80,14 @@ private final class BooleanUserDefaultsDebugViewModel: ObservableObject {
     }
 }
 
-private enum Strings {
-    static let title = NSLocalizedString("debugMenu.booleanUserDefaults.title", value: "Boolean User Defaults", comment: "Boolean User Defaults Debug Menu screen title")
-    static let otherBooleanUserDefaultsSectionID = "Other"
-}
+typealias BooleanUserDefaultsSections = [String: BooleanUserDefaults]
+typealias BooleanUserDefaults = [String: BooleanUserDefault]
 
-private struct BooleanUserDefault {
+struct BooleanUserDefault {
     var title: String
     var value: Bool
+}
+
+private enum Strings {
+    static let otherBooleanUserDefaultsSectionID = "Other"
 }
